% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R, R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday function to find probabilities}
\usage{
birthday(x)

birthday(x)
}
\arguments{
\item{x}{integer, group size(s)}
}
\value{
numeric probability in [0,1]

numeric probability in [0,1]
}
\description{
Computes the probability that in a group of size `n`, at least two people
share a birthday, assuming 365 equally likely birthdays.

Computes the probability that in a group of size `n`, at least two people
share a birthday, assuming 365 equally likely birthdays.
}
\examples{
birthday(20:24)
birthday(20:24)
}
