% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycurve.R
\name{mycurve}
\alias{mycurve}
\title{Draws and shades a normal curve region, returning probability}
\usage{
mycurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Value up to which to shade and calculate P(X ≤ a)}
}
\value{
A list containing mu, sigma, and the probability
}
\description{
Draws and shades a normal curve region, returning probability
}
\examples{
mycurve(mu = 10, sigma = 5, a = 6)

}
